+++ backend/footprint/src/main/java/foot/footprint/domain/article/application/delete/DeleteArticleService.java
@@ -1,0 +1,6 @@
+package foot.footprint.domain.article.application.delete;
+
+public interface DeleteArticleService {
+
+    public void delete(Long articleId, Long memberId);
+}
